Rename DomainSearchControls module files
