+++ packages/domain-search/src/components/domain-search-controls/input.tsx
