+++ packages/domain-search/src/index.ts
@@ -1,5 +1,5 @@
 export { DomainSearch } from './components/DomainSearch/DomainSearch';
-export { DomainSearchControls } from './components/DomainSearchControls/DomainSearchControls';
+export { DomainSearchControls } from './components/domain-search-controls/domain-search-controls';
 export { DomainsMiniCart } from './components/DomainsMiniCart/DomainsMiniCart';
 export { DomainsFullCart } from './components/domains-full-cart';
 export * as DomainSuggestions from './components/DomainSuggestions';
